reviewbot_mgr/requirements.txt
Flask==2.3.3
google==3.0.0
grpcio==1.58.0
grpcio-tools==1.58.0
